I changed the typecast from long to int, but forgot the datatypes in the argument list of the function.

+++ liboct/oct_f.c
@@ -183,7 +183,7 @@ double FC_FUNC_(oct_clock, OCT_CLOCK)
 }
 
 void FC_FUNC_(oct_gettimeofday, OCT_GETTIMEOFDAY)
-	(long *sec, long *usec)
+	(int *sec, int *usec)
 {
 #ifdef linux
   struct timeval tv;
